% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_funcs.R
\name{getMain}
\alias{getMain}
\title{Preprocesses raw output file from an online experiment}
\usage{
getMain(file, sub = NA, rats_all, jstats = T, exp = 1)
}
\arguments{
\item{file}{is a .csv datafile (raw experiment output from the online task)}

\item{sub}{is the specific participant ID (default=NA, when iterating though all files)}

\item{rats_all}{is the rating dataframe created with \code{getRats()}}

\item{jstats}{(bool) whether to include judgement difficulty & magnitude}

\item{exp}{is the experiment number}
}
\description{
Takes in a raw data .csv file (as generated by JavaScript experimental code) as well as the preprocessed ratings and outputs a dataframe, where each row corresponds to a single experimental trial. Creates a set of additional columns with relevant measures.
}
