% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{getRatDiffs}
\alias{getRatDiffs}
\title{Creates a rating df containing choice info}
\usage{
getRatDiffs(main, ratings)
}
\arguments{
\item{main}{a list of main dataframes of all online experiments OR a dataframe from a single experiment}

\item{ratings}{a list of rating dataframes of all online experiments OR a dataframe from a single experiment}
}
\value{
a a rating df with choice counts
}
\description{
Creates a rating df containing choice info
}
