% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{getRatPrePostCor}
\alias{getRatPrePostCor}
\title{Creates a table summarizing session-level correlations between pre and post-task ratings}
\usage{
getRatPrePostCor(ratings, exp)
}
\arguments{
\item{ratings}{ratings data frame}

\item{exp}{experiment number}
}
\value{
a summary tibble of pre and post-task correlations
}
\description{
Creates a table summarizing session-level correlations between pre and post-task ratings
}
