% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_funcs.R
\name{getRats}
\alias{getRats}
\title{Preprocesses participant ratings
Takes in a raw data .csv file (as generated by JavaScript experimental code) and outputs a rating summary table, where each row corresponds to a single item}
\usage{
getRats(file, sub = NA, exp = -1, full_only = F)
}
\arguments{
\item{file}{is a .csv datafile (raw experiment output from the online task)}

\item{sub}{is the specific participant ID (default=NA, when iterating though all files)}

\item{exp}{is the experiment number}

\item{full_only}{(bool) a check whether all files contain 3 full ratings (stops otherwise)}
}
\value{
a data.table with session ratings
}
\description{
Contains functions for preprocessing the raw data
\code{getRatings()} creates a dataframe containing pre and post-task ratings from a single file (session)
\code{getMain()} creates a dataframe containing results of the main part of the experiment
}
