% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{getSizeAcc}
\alias{getSizeAcc}
\title{Creates a table summarizing session-level correlations between size ratings and actual item sizes}
\usage{
getSizeAcc(ratings, exp, method = "spearman")
}
\arguments{
\item{ratings}{data frame containing ratings for}

\item{exp}{experiment number}

\item{method}{correlation method (used by the default r cor.test function)}
}
\value{
a summary tibble with correlations between pre and post-task ratings
}
\description{
Creates a table summarizing session-level correlations between size ratings and actual item sizes
}
