% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_funcs.R
\name{sumRCcons}
\alias{sumRCcons}
\title{Creates a table summarizing session-level correlations between ratings and choices}
\usage{
sumRCcons(data, exp, round_pval = 4)
}
\arguments{
\item{data}{main data frame}

\item{exp}{experiment number}

\item{round_pval}{p-value rounding (how many decimals)}
}
\value{
a summary tibble of rating-choice consistency \%
}
\description{
Creates a table summarizing session-level correlations between ratings and choices
}
